Some components moved in diferent folders
